Add ctrl parameter write_nn
If write_nn is true then we pickle the neural network to disk. This
saves our work because the neural network is expensive to train. It's
unclear how we would use the trained neural network so write_nn is false
by default. Also, torch version 1.5.0 emits a warning that says to use
torch.save instead of pickle.dump. The warning was later removed:

https://github.com/pytorch/pytorch/issues/38597