Use torch 1.4.0
The dgx2 system at osu uses cuda version 10.1. The pypi binary for the
latest torch release, 1.6.0, was compiled with cuda 10.2. I decreased
the torch version until 1.4.0, which was compiled with cuda 10.1

  1.6.0: 10.2
  1.5.1: 10.2
  1.5.0: 10.2
  1.4.0: 10.1

I ran nvidia-smi to determine the cuda version on the osu dgx2. I used

  python3 -c 'import torch; print(torch.version.cuda)'

to determine which cuda version was used to compile torch.