Improve MC particle weighting
Change sourcing and flux tally to use new weighting and delete
num_physical_particles ctrl parameter.

Sourcing
--------
Determine the number of particles sourced in a source region by the
relative length-weighted magnitude (lwm) of the source. For example, if
num_particles is 100

  Key:
    mag    = magnitude of source
    len    = length (spatial extent) of source
    lwm    = mag * len = length-weight magnitude
    parts  = number of particles sourced at this source
    weight = lwm / parts = the weight of each particle sourced

       mag len lwm parts weight
  src1 1.0 1.0 1.0 25    1/25
  src2 3.0 1.0 3.0 75    3/75

Above, len_src1==len_src2 so mag==lwm. Consider 2 sources of unequal len
(taken from reedp1.yaml)

       mag len lwm parts weight
  src1 50  2.0 100 90    100/90
  src2 1.0 1.0 1.0 10    1/10

The key is that the sum of the weights of the particles sourced is equal
to the magnitude of the source.

Why lwm? I'm not sure. If I extend full_slab from 1cm to 2cm I get then
my flux is 1/2 of the analytic flux. If I multiply by the length I get
the correct answer. It might have to do with my flux tally algorithm.

Flux tally
----------
I employ a pathlength tally estimator for the angular flux psi

  psi = (distances * weights) / lengths

where each quantity is a vector of length equal to the number of
particles still alive

  distances contains the pathlengths traveled by the living particles
  in this iteration

  weights contains weights of the living particles

  lengths contains the lengths of the cells in which the particles are
  traveling

  psi contains the angular flux contributions of the living particles in
  this iteration. psi is added to a vector of accumulators which
  accumulate the positive and negative current in each cell. The scalar
  flux is determined by adding the two currents when tracking finishes.