Improve cross section notation
In MC set the scatter probability to be sigma_s0 / sigma_t, not
sigma_s0 / (sigma_s0 + sigma_t). The sum in the latter should have been
sigma_s0 + sigma_a, which is sigma_t.

In the full slab analytic solution use sigma_a, not sigma_t, because the
solution is only valid for a purely absorbing medium.