Use tolerance in regression test fp comparison
Upgrading numerical libraries like numpy and torch triggered failures
due to changes in the flux that were on the order of 1e-7 for nn and
1e-15 for sn and mc (the difference for sn and mc is smaller because
they calculate a 64-bit flux instead of 32-bit). These changes are due
to inconsequential modifications to the fp arithmetic in the libraries
which we don't care about. We could just rebaseline, but then the tests
would fail again the next time we upgrade, and they would still fail now
if someone runs narrows using different versions of numpy and torch.
Do not rebaseline. Use a fuzzy fp comparison instead.