Add spatial loss distribution plot
To plot the flux with the spatial distribution of the loss, run

./analyze.py problem -q fluxloss

Unlike the scalar loss, which we record at every iteration, we only
record the spatial distribution of the loss at convergence.

Also add sharp_slab.yaml, a 2 region problem in which a strong,
pure-absorber with a strong source is next to a void. This resembles the
first 2 regions in the reed problem.

While the sharp_slab mc solution is a nice z-shape (but with a sharp
vertical line instead of a diagonal line connecting the horizontal lines
in the letter z), the nn solution is a gentle slope that descends from
the top horizontal to the lower horizontal. The spatial distribution of
the loss shows that the gentle slope, which is wrong, is also low-loss.
It should be high-loss! This indicates that the calculation of the loss
function is buggy, which makes sense since we just modified it to add
heterogeneity.