Add gpu test
The test does not check the solution against a baseline; it simply
checks that we can run gpu=True, which trains the nn on the gpu. The
test only runs if a gpu is detected and the test makes sure that the
code does not print 'Skipping gpu' which is what happens when the user
requests gpu training with gpu=True but the code does not detect a gpu.