Add max_num_src_iter and repurpose max_num_iter
max_num_iter is the maximum number of iterations we will train for
before we quit trying to converge. Use this in combination with epsilon
to run a fixed number of iterations.

max_num_src_iter does what max_num_iter used to do; it's the maximum
number of iterations we will allow in the sn solver source iteration
process before we quit trying to converge the source.